requirements-dev.txt
django-debug-toolbar==3.7.0
wagtail-factories==3.1.0
coverage==6.5.0
